types: uint64 in heights, part 2

+++ blockchain/reactor.go
@@ -321,7 +321,7 @@ func DecodeMessage(bz []byte, maxSize int) (msgType byte, msg BlockchainMessage,
 //-------------------------------------
 
 type bcBlockRequestMessage struct {
-	Height int
+	Height uint64
 }
 
 func (m *bcBlockRequestMessage) String() string {
@@ -329,7 +329,7 @@ func (m *bcBlockRequestMessage) String() string {
 }
 
 type bcNoBlockResponseMessage struct {
-	Height int
+	Height uint64
 }
 
 func (brm *bcNoBlockResponseMessage) String() string {
@@ -350,7 +350,7 @@ func (m *bcBlockResponseMessage) String() string {
 //-------------------------------------
 
 type bcStatusRequestMessage struct {
-	Height int
+	Height uint64
 }
 
 func (m *bcStatusRequestMessage) String() string {
@@ -360,7 +360,7 @@ func (m *bcStatusRequestMessage) String() string {
 //-------------------------------------
 
 type bcStatusResponseMessage struct {
-	Height int
+	Height uint64
 }
 
 func (m *bcStatusResponseMessage) String() string {
