+++ state/errors.go
@@ -9,13 +9,13 @@ type (
 	ErrProxyAppConn error
 
 	ErrUnknownBlock struct {
-		Height int
+		Height uint64
 	}
 
 	ErrBlockHashMismatch struct {
 		CoreHash []byte
 		AppHash  []byte
-		Height   int
+		Height   uint64
 	}
 
 	ErrAppBlockHeightTooHigh struct {
@@ -24,7 +24,7 @@ type (
 	}
 
 	ErrLastStateMismatch struct {
-		Height int
+		Height uint64
 		Core   []byte
 		App    []byte
 	}
@@ -35,7 +35,7 @@ type (
 	}
 
 	ErrNoValSetForHeight struct {
-		Height int
+		Height uint64
 	}
 )
